Functions for license files
